+++ static/output.txt
@@ -1,0 +1,48 @@
+2018-12-15 20:11
+Why ‘Grieve’ Is the Word of the Year
+Link is -->   https://medium.com/p/1662e2fa4941?source=topic_page---------------------------20
+
+2018-12-15 20:11
+What we learned about Diversity & Inclusion in 2018…
+Link is -->   https://medium.com/p/d2f1cefcdf7f?source=topic_page---------0------------------1
+
+2018-12-15 20:11
+The City at the Center of the Cosmos
+Link is -->   https://medium.com/p/11b3b6db57c9?source=topic_page---------1------------------1
+
+2018-12-15 20:11
+When You Listen to Music, You’re Never Alone
+Link is -->   https://medium.com/p/1845e096986?source=topic_page---------2------------------1
+
+2018-12-15 20:11
+Alexander Chee
+Link is -->   https://medium.com/@alexanderchee
+
+2018-12-15 20:11
+The Life and Death of Rap’s Leak Era
+Link is -->   https://medium.com/p/ad00eff519e6?source=topic_page---------4------------------1
+
+2018-12-15 20:11
+Why ‘Grieve’ Is the Word of the Year
+Link is -->   https://medium.com/p/1662e2fa4941?source=topic_page---------5------------------1
+
+2018-12-15 20:11
+A Year of Rowling Dangerously
+Link is -->   https://medium.com/p/68f7c38095d3?source=topic_page---------6------------------1
+
+2018-12-15 20:11
+The Blue Square of Homelessness Continues!
+Link is -->   https://medium.com/p/a09bd187e340?source=topic_page---------7------------------1
+
+2018-12-15 20:11
+Sex Work & Stigma: Examining the Problematic Representation of Sex Workers in Popular Culture
+Link is -->   https://medium.com/p/5f0e875d4aac?source=topic_page---------8------------------1
+
+2018-12-15 20:11
+My (and Many Other Black Women’s) Complicated Relationship with Rap
+Link is -->   https://medium.com/p/92b110234954?source=topic_page---------9------------------1
+
+2018-12-15 20:11
+Are You Cyberpunk Enough Now to Unplug from the Mainstream?
+Link is -->   https://medium.com/p/98080604e742?source=topic_page---------10------------------1
+
