--- templates/download.html
@@ -1,9 +1,0 @@
-{% extends "layout.html" %}
-
-{% block heading %}
-	FINISHED! Please check the output file
-{% endblock %}
-
-{% block body %}
-	
-{% endblock %}
