+++ templates/index.html
@@ -12,15 +12,6 @@
 				<form class="form-horizontal" action="{{url_for('result')}}" method="post"> 
 					<div class="form-group">
 						<div class="col-sm-12">
-							<input class="form-control" type="text" list="browsers" name="browser" placeholder="Choose the browser:">
-							<datalist id="browsers">
-							    <option value="Firefox">
-							    <option value="Chrome">
-							</datalist>
-						</div>
-					</div>
-					<div class="form-group">
-						<div class="col-sm-12">
 							<input class="form-control" type="text" list="topics" name="topic" placeholder="Where would you like to dive in?">
 							<datalist id="topics">
 							    <option value="Home">
